moved retry logic to separated file and separated package - also possible to set number of reties and duration of sleep
